# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de gestion des personnes",
        "description": "Documentation de l'API pour gérer les personnes",
        "version": "1.0.0"
    },
    "paths": {
        "/api/personnes": {
            "get": {
                "tags": [
                    "Personnes"
                ],
                "summary": "Récupérer toutes les personnes",
                "operationId": "a98e7cdd225a906ffd217cba4430558b",
                "responses": {
                    "200": {
                        "description": "Liste des personnes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Personne"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Personnes"
                ],
                "summary": "Créer une nouvelle personne",
                "operationId": "37c59a8f658db88ec1950a749bfb83a6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Personne"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Personne créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Personne"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/personnes/{id}": {
            "get": {
                "tags": [
                    "Personnes"
                ],
                "summary": "Récupérer une personne par ID",
                "operationId": "e0a8cbd54a6a341687fe44c01f1c938f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la personne",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la personne",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Personne"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Personne non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "Personnes"
                ],
                "summary": "Mettre à jour une personne",
                "operationId": "12bceb6a49b10be6fc7329efb78290f7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la personne à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Personne"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Personne mise à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Personne"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Personne non trouvée"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Personnes"
                ],
                "summary": "Supprimer une personne",
                "operationId": "2687ddb9d8cb54ea6137605c582db2de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la personne à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Personne supprimée avec succès"
                    },
                    "404": {
                        "description": "Personne non trouvée"
                    }
                }
            }
        },
        "/api/projets": {
            "get": {
                "tags": [
                    "Projets"
                ],
                "summary": "Récupérer tous les projets",
                "operationId": "cdcc7d8fa82f1533384f69fd6bf52377",
                "responses": {
                    "200": {
                        "description": "Liste des projets",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Projet"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Projets"
                ],
                "summary": "Créer un nouveau projet",
                "operationId": "b09251428bbf1995144e555a6db447ae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Projet"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Projet créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Projet"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/projets/{id}": {
            "get": {
                "tags": [
                    "Projets"
                ],
                "summary": "Récupérer un projet par ID",
                "operationId": "c12153e49ba7b2785975d94224588a62",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du projet",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du projet",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Projet"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Projet non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Projets"
                ],
                "summary": "Mettre à jour un projet",
                "operationId": "f93c637a8568fe1980bc6bde69679dca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du projet",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Projet"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Projet mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Projet"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Projet non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Projets"
                ],
                "summary": "Supprimer un projet",
                "operationId": "d880737fcbd65754b670b8e357c18c12",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du projet",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Projet supprimé avec succès"
                    },
                    "404": {
                        "description": "Projet non trouvé"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Personne": {
                "title": "Personne",
                "description": "Modèle représentant une personne",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nom": {
                        "type": "string",
                        "example": "Dupont"
                    },
                    "prenom": {
                        "type": "string",
                        "example": "Jean"
                    },
                    "ville": {
                        "type": "string",
                        "example": "Paris"
                    },
                    "telephone": {
                        "type": "string",
                        "example": "+33123456789"
                    }
                },
                "type": "object"
            },
            "Projet": {
                "title": "Projet",
                "description": "Model representing a project",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "date_de_debut": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-02-25"
                    },
                    "date_de_fin": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-31"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Personnes",
            "description": "Endpoints pour gérer les personnes"
        },
        {
            "name": "Projets",
            "description": "Operations related to projects"
        }
    ]
}